+++ jest-runner-eslint.config.js
@@ -1,0 +1,9 @@
+'use strict';
+
+module.exports = {
+	cliOptions: {
+		cache: true,
+		quiet: true,
+		reportUnusedDisableDirectives: 'error'
+	}
+};